% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-theme.R
\name{gt_theme_measr}
\alias{gt_theme_measr}
\title{Personalized gt theme}
\usage{
gt_theme_measr(gt_dat, bg_color = "#FFFFFF", ...)
}
\arguments{
\item{gt_dat}{A \code{gt_tbl} object, generated from \link[gt:gt]{gt::gt}.}

\item{bg_color}{The background color of the table cells.}

\item{...}{Additional arguments passed to \link[gt:tab_options]{gt::tab_options}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
A custom theme for tables generated with \link[gt:gt]{gt::gt}.
}
