% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palette_measr}
\alias{palette_measr}
\title{Color palette based on measr branding}
\format{
An object of class \code{character} of length 5.
}
\usage{
palette_measr
}
\value{
A character vector of colors.
}
\description{
A color palette inspired by the hex sticker logo for the measr package.
}
\keyword{datasets}
