% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-measr.R
\name{theme_measr}
\alias{theme_measr}
\title{Title}
\usage{
theme_measr(base_family = "Open Sans", base_size = 11.5, ...)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{...}{Additional arguments passed to \link[hrbrthemes:theme_ipsum]{hrbrthemes::theme_ipsum}.}
}
\value{
A ggplot2 theme.
}
\description{
An opinionated ggplot2 theme based on \code{\link[hrbrthemes:theme_ipsum]{hrbrthemes::theme_ipsum()}}. The theme
automatically supports Markdown in axis and facet labels through
\code{\link[ggtext:element_markdown]{ggtext::element_markdown()}} and provides reasonable defaults for visually
appealing graphics.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = disp)) +
  geom_point() +
  theme_measr()
}
